Conflict

+++ Assets/Mana/GreenMana.cs
@@ -1,14 +1,14 @@
 ﻿public class GreenMana
 {
-    public static void BurnGreen(float amount)
+    public static void BurnGreen(int amount)
     {
     }
 
-    public static void GainGreen(float amount)
+    public static void GainGreen(int amount)
     {
     }
 
-    public static Mana GetBase(float startingTotal = 0)
+    public static Mana GetBase(int startingTotal = 0)
     {
         return new Mana(ManaColor.Green, GainGreen, BurnGreen)
         {
