+++ Assets/Mana/Mana.cs
@@ -3,26 +3,26 @@ using UnityEngine;
 
 public class Mana
 {
-    public Mana(ManaColor color, Action<float> gainAction, Action<float> burnAction)
+    public Mana(ManaColor color, Action<int> gainAction, Action<int> burnAction)
     {
         GainAction = gainAction;
         BurnAction = burnAction;
         Color = color;
     }
 
-    public Action<float> BurnAction { get; set; }
+    public Action<int> BurnAction { get; set; }
     public ManaColor Color { get; set; }
-    public Action<float> GainAction { get; set; }
+    public Action<int> GainAction { get; set; }
 
-    public float Total { get; set; }
+    public int Total { get; set; }
 
-    public void Burn(float amount)
+    public void Burn(int amount)
     {
         Total -= amount;
         BurnAction?.Invoke(amount);
     }
 
-    public void Gain(float amount)
+    public void Gain(int amount)
     {
         Total += amount;
         GainAction?.Invoke(amount);
