+++ Assets/Mana/WhiteMana.cs
@@ -1,14 +1,14 @@
 ﻿public class WhiteMana
 {
-    public static void BurnWhite(float amount)
+    public static void BurnWhite(int amount)
     {
     }
 
-    public static void GainWhite(float amount)
+    public static void GainWhite(int amount)
     {
     }
 
-    public static Mana GetBase(float startingTotal = 0)
+    public static Mana GetBase(int startingTotal = 0)
     {
         return new Mana(ManaColor.White, GainWhite, BurnWhite)
         {
